ctvt_api/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
bcrypt
python-jose
